refactor: move settings_store.py and file_settings_store.py to app_server/settings
Move settings store implementations from openhands/storage/settings to
openhands/app_server/settings to follow the V1 convention of packaging
by feature rather than by layer.

Updated all imports across the codebase and enterprise folder.
